+++ src/client/event.ts
@@ -1,4 +1,4 @@
-import { PublishEventEnvelope } from "../protobuf/clientchannel/event_pb";
+import { PublishEventEnvelope } from "@fraym/streams-proto/dist/event";
 
 export interface SubscriptionEvent extends BaseEvent {
     topic: string;
@@ -32,34 +32,38 @@ export type HandlerFunc = (event: SubscriptionEvent) => Promise<void>;
 export const getSubscriptionEvent = (
     eventEnvelope: PublishEventEnvelope
 ): SubscriptionEvent | null => {
-    const event = eventEnvelope.getEvent();
+    const event = eventEnvelope.event;
     if (!event) {
         return null;
     }
 
     const payload: Record<string, EventData> = {};
 
-    event.getPayloadMap().forEach((data, key) => {
-        if (data.hasGdpr()) {
-            payload[key] = {
-                value: data.getValue(),
-                gdprDefault: data.getGdpr()?.getDefault() ?? "",
-            };
-        } else {
-            payload[key] = data.getValue();
+    for (const key in event.payload) {
+        if (Object.prototype.hasOwnProperty.call(event.payload, key)) {
+            const data = event.payload[key];
+
+            if (data.metadata && data.metadata.gdpr) {
+                payload[key] = {
+                    value: data.value,
+                    gdprDefault: data.metadata.gdpr.default ?? "",
+                };
+            } else {
+                payload[key] = data.value;
+            }
         }
-    });
+    }
 
     return {
-        id: event.getId(),
-        topic: eventEnvelope.getTopic(),
-        tenantId: eventEnvelope.getTenantId(),
+        id: event.id,
+        topic: eventEnvelope.topic,
+        tenantId: eventEnvelope.tenantId,
         payload,
-        raisedAt: new Date(event.getRaisedAt()),
-        stream: event.getStream() || undefined,
-        type: event.getType() || undefined,
-        causationId: event.getCausationId() || undefined,
-        correlationId: event.getCorrelationId() || undefined,
-        reason: event.getReason() || undefined,
+        raisedAt: new Date(event.raisedAt),
+        stream: event.stream || undefined,
+        type: event.type || undefined,
+        causationId: event.causationId || undefined,
+        correlationId: event.correlationId || undefined,
+        reason: event.reason || undefined,
     };
 };