+++ src/client/handler.ts
@@ -1,7 +1,7 @@
 import { Stream } from "./init";
-import { Response } from "../protobuf/clientchannel/response_pb";
 import { getSubscriptionEvent, HandlerFunc, SubscriptionEvent } from "./event";
-import { Request } from "../protobuf/clientchannel/request_pb";
+import { Response } from "@fraym/streams-proto/dist/response";
+import { Request } from "@fraym/streams-proto/dist/request";
 
 interface EventHandler {
     addHandler: (type: string, handler: HandlerFunc) => void;
@@ -13,8 +13,11 @@ export const initEventHandler = (stream: Stream): EventHandler => {
     const globalHandlers: HandlerFunc[] = [];
 
     stream.on("data", (data: Response) => {
-        const eventEnvelope = data.hasEvent() ? data.getEvent() : null;
-        const event = eventEnvelope ? getSubscriptionEvent(eventEnvelope) : null;
+        if (data.data?.$case !== "event") {
+            return;
+        }
+
+        const event = getSubscriptionEvent(data.data.event);
         if (!event) {
             return;
         }
@@ -49,14 +52,16 @@ const handleEvent = (event: SubscriptionEvent, handler: HandlerFunc, stream: Str
 };
 
 const newEventAckRequest = (tenantId: string, topic: string, eventId: string): Request => {
-    const action = new Request.EventAck();
-    action.setTenantId(tenantId);
-    action.setTopic(topic);
-    action.setEventId(eventId);
-
-    const request = new Request();
-    request.setEventAck(action);
-    return request;
+    return {
+        payload: {
+            $case: "eventAck",
+            eventAck: {
+                eventId,
+                tenantId,
+                topic,
+            },
+        },
+    };
 };
 
 const newEventNotAckRequest = (
@@ -65,13 +70,15 @@ const newEventNotAckRequest = (
     eventId: string,
     reason: string
 ): Request => {
-    const action = new Request.EventNotAck();
-    action.setTenantId(tenantId);
-    action.setTopic(topic);
-    action.setEventId(eventId);
-    action.setReason(reason);
-
-    const request = new Request();
-    request.setEventNotAck(action);
-    return request;
+    return {
+        payload: {
+            $case: "eventNotAck",
+            eventNotAck: {
+                eventId,
+                tenantId,
+                topic,
+                reason,
+            },
+        },
+    };
 };